Move `FUSION_SOURCE_EXTENSION` into `SourceName`.
It's reasonable for this to be API.
