+++ runtime/src/main/java/dev/ionfusion/fusion/FileSystemModuleRepository.java
@@ -3,8 +3,8 @@
 
 package dev.ionfusion.fusion;
 
-import static dev.ionfusion.fusion.GlobalState.FUSION_SOURCE_EXTENSION;
 import static dev.ionfusion.fusion.ModuleIdentity.isValidModuleName;
+import static dev.ionfusion.fusion.SourceName.FUSION_SOURCE_EXTENSION;
 
 import java.io.File;
 import java.util.function.Consumer;
